Fixed a bug in frame's move assignment operator.
The zmq_msg_init() at the beginning created a new message overwriting
the handle of the old message. But this old message was not closed
leaving a memory leak behind. Instead of closing the existing handle
explicitly we can simply remove the init call because zmq_msg_move()
does the closing of the old message on its own.

see https://github.com/zeromq/zmqpp/issues/204